refactor: moved base GP linear Gaussian transition models from base.py to linear.py
